Simplifying project structure and adding AMD support
* Removing src folder and file in favor of full structure defined in
index.js where support for AMD (require.js) has been added. Library now
supports a universal pattern for module loading (AMD, CommonJS, and
traditional).
* Removed grunt packages and tasks that were not being used.
