+++ test/karma.conf.js
@@ -9,7 +9,7 @@ module.exports = function(config) {
     files: [
       'bower_components/angular/angular.js',
       'bower_components/angular-mocks/angular-mocks.js',
-      'src/*.js',
+      'index.js',
       'test/unit/*.js'
     ],
     reporters: ['progress'],
